Changes in the ui design: main and detail

+++ app/src/main/java/com/example/mvvmpokedex/model/PokemonModel.kt
@@ -1,3 +1,3 @@
 package com.example.mvvmpokedex.model
 
-data class PokemonModel(val name: String, val height: Int, val weight: Int, val image: String)
+data class PokemonModel(val name: String, val height: Double, val weight: Double, val image: String)
